feat: move Event service into a event folder
